requirement.txt
fastapi

flask

tensorflow

numpy

pillow

requests

flask-cors

werkzeug